+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.skiapp",
+    package = "com.sbxapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.skiapp",
+    package = "com.sbxapp",
     res = "src/main/res",
 )
