+++ android/app/src/main/java/com/sbxapp/MainApplication.java
@@ -1,9 +1,8 @@
-package com.skiapp;
+package com.sbxapp;
 
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.lwansbrough.RCTCamera.RCTCameraPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -23,8 +22,7 @@ public class MainApplication extends Application implements ReactApplication {
     @Override
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-            new RCTCameraPackage()
+          new MainReactPackage()
       );
     }
