+++ android/settings.gradle
@@ -1,5 +1,3 @@
-rootProject.name = 'SkiApp'
-include ':react-native-camera'
-project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-camera/android')
+rootProject.name = 'SBXApp'
 
 include ':app'
