+++ app.json
@@ -2,6 +2,6 @@
   "expo": {
     "sdkVersion": "22.0.0"
   },
-  "name": "SkiApp",
-  "displayName": "SkiApp"
+  "name": "SBXApp",
+  "displayName": "SBX App"
 }
