+++ index.js
@@ -1,3 +1,3 @@
 import { AppRegistry } from 'react-native';
 import AppNavigator from './Navigators/AppNavigator';
-AppRegistry.registerComponent('SkiApp', () => AppNavigator);
+AppRegistry.registerComponent('SBXApp', () => AppNavigator);
