+++ ios/SBXApp.xcodeproj/xcshareddata/xcschemes/SBXApp-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "SkiApp-tvOS.app"
-               BlueprintName = "SkiApp-tvOS"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXApp-tvOS.app"
+               BlueprintName = "SBXApp-tvOS"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "SkiApp-tvOSTests.xctest"
-               BlueprintName = "SkiApp-tvOSTests"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXApp-tvOSTests.xctest"
+               BlueprintName = "SBXApp-tvOSTests"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "SkiApp-tvOSTests.xctest"
-               BlueprintName = "SkiApp-tvOSTests"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXApp-tvOSTests.xctest"
+               BlueprintName = "SBXApp-tvOSTests"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "SkiApp-tvOS.app"
-            BlueprintName = "SkiApp-tvOS"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp-tvOS.app"
+            BlueprintName = "SBXApp-tvOS"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "SkiApp-tvOS.app"
-            BlueprintName = "SkiApp-tvOS"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp-tvOS.app"
+            BlueprintName = "SBXApp-tvOS"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "SkiApp-tvOS.app"
-            BlueprintName = "SkiApp-tvOS"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp-tvOS.app"
+            BlueprintName = "SBXApp-tvOS"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
