+++ ios/SBXApp.xcodeproj/xcshareddata/xcschemes/SBXApp.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "SkiApp.app"
-               BlueprintName = "SkiApp"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXApp.app"
+               BlueprintName = "SBXApp"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "SkiAppTests.xctest"
-               BlueprintName = "SkiAppTests"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXAppTests.xctest"
+               BlueprintName = "SBXAppTests"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "SkiAppTests.xctest"
-               BlueprintName = "SkiAppTests"
-               ReferencedContainer = "container:SkiApp.xcodeproj">
+               BuildableName = "SBXAppTests.xctest"
+               BlueprintName = "SBXAppTests"
+               ReferencedContainer = "container:SBXApp.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "SkiApp.app"
-            BlueprintName = "SkiApp"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp.app"
+            BlueprintName = "SBXApp"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "SkiApp.app"
-            BlueprintName = "SkiApp"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp.app"
+            BlueprintName = "SBXApp"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "SkiApp.app"
-            BlueprintName = "SkiApp"
-            ReferencedContainer = "container:SkiApp.xcodeproj">
+            BuildableName = "SBXApp.app"
+            BlueprintName = "SBXApp"
+            ReferencedContainer = "container:SBXApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
