+++ ios/SBXApp/Images.xcassets/AppIcon.appiconset/Contents.json
