+++ ios/SBXAppTests/Info.plist
