+++ ios/SBXAppTests/SBXAppTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface SkiAppTests : XCTestCase
+@interface SBXAppTests : XCTestCase
 
 @end
 
-@implementation SkiAppTests
+@implementation SBXAppTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
