+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "SkiApp",
+  "name": "SBXApp",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
